pachacutin_ai/requirements.txt
flask
torch
torchvision
opencv-python
pillow
pyserial
pytest
openai
